Initial layout for TeamPay.
This is just the sceleton for the app. It contains the basic
controllers, models and views to create games and collect payment.

It also integrates with Google API for User Auth.

It still lacks administrator and player roles and their respective
views.

All the sensitive data is loaded into the ENV params.